Add pause/resume and update when necessary
Updating of file may be manually paused and resumed using the very crude
links next to the reload delay input.

A file will only be reloaded if its modification date has changed. This
allows us to inspect the html in a debugger, as well as scroll in
overflowing block elements. Without this the file would continually
reload, resetting scroll and content, which prevents us from properly
inspecting the file.